Worked on the GUI + added the possibility to load an image from an url

+++ dante/HttpRequest/MP3Downloader.cs
@@ -3,7 +3,7 @@ namespace HttpRequest
 {
     public class MP3Downloader
     {
-        public async static void DownloadSong(int id, string destination)
+        public async static void DownloadSong(long id, string destination)
         {
             HttpClient client = new HttpClient();
             HttpResponseMessage response = await client.GetAsync($"https://dante.kartof.tk/song/{id}", HttpCompletionOption.ResponseHeadersRead);
